refactor(modules): 将 AutoScriptEngine 移至 dsl/，ConfigUtils 移至 managers/config/，修复单一职责和依赖倒置问题
