+++ src/dsl/__init__.py
@@ -1,0 +1,9 @@
+"""
+    DSL module for the AutoLibrary project.
+
+    Contains the AutoScript DSL engine and related components
+    for preprocessing user reservation data in timer tasks.
+
+    Classes:
+    - AutoScriptEngine: AutoScript script engine class.
+"""
