+++ src/utils/__init__.py
@@ -5,6 +5,4 @@
     - TimerUtils: Timer utils class for the AutoLibrary project.
     - JSONReader: JSON reader class for the AutoLibrary project.
     - JSONWriter: JSON writer class for the AutoLibrary project.
-    - ConfigUtils: Config utils class for the AutoLibrary project.
-    - AutoScriptEngine: AutoScript script engine class for the AutoLibrary project.
 """
